Uploaded the supporting document
In this document, you will find a brief explanation about what I have done in this repository and relating videos. The video explanations were done for each of the folders (Theoretical approach, 2 Practical approaches and the extra explanation) and you can find those links in this document at each relevant section dedicated to the approaches.